+++ endpoint_tsc/src/chisel-global.d.ts
@@ -1,0 +1,4 @@
+import * as ChiselAlias from "@chiselstrike/api";
+declare global {
+    let Chisel: typeof ChiselAlias;
+}
