+++ my_tsc/src/main.rs
@@ -1,7 +1,7 @@
 use deno_runtime::inspector_server::InspectorServer;
+use endpoint_tsc::Compiler;
 use std::net::SocketAddr;
 use structopt::StructOpt;
-use tsc_compile::Compiler;
 
 #[derive(Debug, StructOpt)]
 struct Opt {
@@ -20,16 +20,14 @@ async fn main() {
     if opt.inspect {
         let addr: SocketAddr = "127.0.0.1:9229".parse().unwrap();
         let inspector = InspectorServer::new(addr, "my_tsc".to_string());
-        inspector.register_inspector("main_module".to_string(), &mut compiler.runtime, true);
+        inspector.register_inspector("main_module".to_string(), &mut compiler.tsc.runtime, true);
         _maybe_inspector = Some(inspector);
         compiler
+            .tsc
             .runtime
             .inspector()
             .wait_for_session_and_break_on_next_statement();
     }
 
-    compiler
-        .compile_ts_code(&opt.file, Default::default())
-        .await
-        .unwrap();
+    compiler.compile_endpoint(&opt.file).await.unwrap();
 }
